+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Added `Logger` class.
 - Added `CliOptions` for simple command-line parsing.
 - Added public function to encapsulate use of nlohmann/json and magic_enum from users.
+- Added complete basic consumer example project used for testing installation
 
 ## v0.1